+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Release Notes for Craft Commerce
 
-## 2.1.13 - 2019-09-09
+## Unreleased - 2019-09-09
+
+### Added
+- Added `craft\commerce\controllers\BaseFrontEndController::EVENT_MODIFY_CART_CONTENT`. ([#1002](https://github.com/craftcms/commerce/issues/1002))
 
 ### Fixed
 - Fixed a error when requesting a PDF URL in headless mode. ([#1011](https://github.com/craftcms/commerce/pull/1011))
+- `Order::datePaid` is now set to null if the order is no longer paid iun full after a refund. ([#1026](https://github.com/craftcms/commerce/pull/1026))
 
 ## 2.1.13 - 2019-09-09